# artifacts/afucloud-backend/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: AfuCloud API – cloud platform for storing, managing, and delivering digital assets
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all user projects
      responses:
        "200":
          description: Project list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /v1/projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get project by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Project updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /v1/projects/{id}/stats:
    get:
      operationId: getProjectStats
      tags: [projects]
      summary: Get storage and usage stats for a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Project statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectStats"

  # ─── IMAGES ───
...
  /v1/tokens:
    get:
      operationId: listTokens
      tags: [tokens]
      summary: List personal access tokens
      responses:
        "200":
          description: Tokens
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Token"
    post:
      operationId: createToken
      tags: [tokens]
      summary: Create a personal access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TokenInput"
      responses:
        "201":
          description: Token created (secret shown once)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenWithSecret"
  /v1/tokens/{id}:
    delete:
      operationId: revokeToken
      tags: [tokens]
      summary: Revoke a personal access token
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Token revoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ─── ANALYTICS ───
...
  /v1/projects/{projectId}/webhooks:
    get:
      operationId: listWebhooks
      tags: [webhooks]
      summary: List webhooks for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Webhooks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Webhook"
    post:
      operationId: createWebhook
      tags: [webhooks]
      summary: Create a webhook
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebhookInput"
      responses:
        "201":
          description: Webhook created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Webhook"
  /v1/projects/{projectId}/webhooks/{id}:
    patch:
      operationId: updateWebhook
      tags: [webhooks]
      summary: Update a webhook
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebhookUpdate"
      responses:
        "200":
          description: Webhook updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Webhook"
    delete:
      operationId: deleteWebhook
      tags: [webhooks]
      summary: Delete a webhook
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ─── ACTIVITY LOGS ───
...
components:
  schemas:
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    # ─── Auth ───
...
    Project:
      type: object
      required: [id, name, slug, userId, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        userId:
          type: string
        imageCount:
          type: number
        storageUsed:
          type: number
        createdAt:
          type: string
        updatedAt:
          type: ["string", "null"]
    ProjectInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
    ProjectUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
    ProjectStats:
      type: object
      required: [totalImages, storageUsed, bandwidth, apiRequests]
      properties:
        totalImages:
          type: number
        storageUsed:
          type: number
        bandwidth:
          type: number
        apiRequests:
          type: number
        deletedImages:
          type: number
        favoriteImages:
          type: number

    # ─── Images ───
...
    Token:
      type: object
      required: [id, name, prefix, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        prefix:
          type: string
        scopes:
          type: array
          items:
            type: string
        expiresAt:
          type: ["string", "null"]
        lastUsedAt:
          type: ["string", "null"]
        revokedAt:
          type: ["string", "null"]
        createdAt:
          type: string
    TokenInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
        scopes:
          type: array
          items:
            type: string
        expiresAt:
          type: ["string", "null"]
    TokenWithSecret:
      allOf:
        - $ref: "#/components/schemas/Token"
        - type: object
          required: [secret]
          properties:
            secret:
              type: string

    # ─── Analytics ───
...
    Webhook:
      type: object
      required: [id, projectId, url, events, active, createdAt]
      properties:
        id:
          type: string
        projectId:
          type: string
        url:
          type: string
        events:
          type: array
          items:
            type: string
        active:
          type: boolean
        createdAt:
          type: string
    WebhookInput:
      type: object
      required: [url, events]
      properties:
        url:
          type: string
        events:
          type: array
          items:
            type: string
        secret:
          type: string
    WebhookUpdate:
      type: object
      properties:
        url:
          type: string
        events:
          type: array
          items:
            type: string
        active:
          type: boolean
        secret:
          type: ["string", "null"]

    # ─── Activity ───
...